disconnected all database implementations from the adapter (Only mysql left)
